feat: :sparkles: add partial uci support
