+++ src/util.go
@@ -87,102 +87,6 @@ func StringToCoord(s string) (Coord, error) {
 	return coord, nil
 }
 
-func (b *Board) UCIParseMove(s string) (Move, error) {
-	sParts := []rune(s)
-
-	if len(sParts) < 4 || len(sParts) > 5 {
-		return 0, fmt.Errorf("Invalid UCI move notation: %v", s)
-	}
-
-	originCoord, err := StringToCoord(string(sParts[:2]))
-	if err != nil {
-		return 0, err
-	}
-
-	dstCoord, err := StringToCoord(string(sParts[2:4]))
-	if err != nil {
-		return 0, err
-	}
-
-	originSquare := b.squares[originCoord]
-	dstSquare := b.squares[dstCoord]
-
-	moveKind := QUIET
-	if dstSquare != EMPTY {
-		moveKind = CAPTURE
-	}
-
-	if originSquare == WHITE_PAWN && dstCoord-originCoord == 16 {
-		moveKind = DOUBLE_PAWN_PUSH
-	}
-
-	if originSquare == BLACK_PAWN && originCoord-dstCoord == 16 {
-		moveKind = DOUBLE_PAWN_PUSH
-	}
-
-	if dstCoord == b.epCoord && (originSquare == WHITE_PAWN || originSquare == BLACK_PAWN) {
-		moveKind = EP_CAPTURE
-	}
-
-	if originSquare == WHITE_KING {
-		switch dstCoord {
-		case G1:
-			if b.castleRights&CASTLING_RIGHTS_WHITE_KING_MASK > 0 {
-				moveKind = KING_CASTLE
-			}
-		case C1:
-			if b.castleRights&CASTLING_RIGHTS_WHITE_QUEEN_MASK > 0 {
-				moveKind = QUEEN_CASTLE
-			}
-		}
-	}
-
-	if originSquare == BLACK_KING {
-		switch dstCoord {
-		case G8:
-			if b.castleRights&CASTLING_RIGHTS_BLACK_KING_MASK > 0 {
-				moveKind = KING_CASTLE
-			}
-		case C8:
-			if b.castleRights&CASTLING_RIGHTS_BLACK_QUEEN_MASK > 0 {
-				moveKind = QUEEN_CASTLE
-			}
-		}
-	}
-
-	if len(sParts) == 5 {
-		switch sParts[4] {
-		case 'n':
-			if moveKind == CAPTURE {
-				moveKind = KNIGHT_PROMOTION_CAPTURE
-			} else {
-				moveKind = KNIGHT_PROMOTION
-			}
-		case 'b':
-			if moveKind == CAPTURE {
-				moveKind = BISHOP_PROMOTION_CAPTURE
-			} else {
-				moveKind = BISHOP_PROMOTION
-			}
-		case 'r':
-			if moveKind == CAPTURE {
-				moveKind = ROOK_PROMOTION_CAPTURE
-			} else {
-				moveKind = ROOK_PROMOTION
-			}
-		case 'q':
-			if moveKind == CAPTURE {
-				moveKind = QUEEN_PROMOTION_CAPTURE
-			} else {
-				moveKind = QUEEN_PROMOTION
-			}
-		default:
-			return 0, fmt.Errorf("Invalid promotion: %v", s)
-		}
-	}
-	return NewMove(originCoord, dstCoord, originSquare, dstSquare, moveKind), nil
-}
-
 func IntPow(a int, b int) int {
 	result := 1
